AI_Employee_Vault/gold/pyproject.toml
[project]
name = "gold"
version = "3.0.0"
description = "Gold-tier Personal AI Employee — Hackathon"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Gmail Watcher — Google API client stack
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.126.0",
    # LinkedIn + WhatsApp + Facebook Watchers — Playwright browser automation
    "playwright>=1.49.0",
    # Filesystem Watcher
    "watchdog>=4.0.0",
    # Scheduler — daily 8 AM + weekly Sunday cron-like triggers
    "schedule>=1.2.0",
    # Environment variable loading (.env files)
    "python-dotenv>=1.0.0",
    # Twitter API v2 client
    "tweepy>=4.14.0",
    # HTTP client for Odoo JSON-RPC and Meta Graph API
    "httpx>=0.27.0",
    # Retry logic with exponential backoff
    "tenacity>=8.2.0",
]

[tool.uv]
# This project is a collection of scripts, not an installable Python package.
# Disable editable wheel building so `uv sync` only installs dependencies.
package = false
